Move blacklist_from_iter to near-network-primitives"
